@@ -1,8 +1,8 @@
 import copy
 import datetime
 import operator
+from collections.abc import Callable
 from functools import reduce
-from typing import Callable
 
 import numpy
 import pandas as pd
@@ -172,7 +172,6 @@ def _parse_einführungsfaktor_vorsorgeaufw_alter_ab_2005(date, params):
     """
     jahr = float(date.year)
     if jahr >= 2005:
-        # ToDo: remove conversion to Series after moving to scalar
         out = piecewise_polynomial(
             pd.Series(jahr),
             thresholds=params["eink_st_abzuege"]["einführungsfaktor"]["thresholds"],
@@ -238,11 +237,12 @@ def load_functions_for_date(date):
     """
 
     # Using TIME_DEPENDENT_FUNCTIONS here leads to failing tests.
-    functions = {
-        f.__info__["dates_active_dag_key"]: f
-        for f in load_internal_functions().values()
-        if is_time_dependent(f) and is_active_at_date(f, date)
-    }
+    functions = {}
+    for f in load_internal_functions().values():
+        if not is_time_dependent(f) or is_active_at_date(f, date):
+            info = f.__info__ if hasattr(f, "__info__") else {}
+            name = info.get("dates_active_dag_key", f.__name__)
+            functions[name] = f
 
     return functions
 
@@ -287,6 +287,13 @@ def _load_parameter_group_from_yaml(
         # Take care of leap years
         except ValueError:
             dt = dt.replace(year=dt.year - years, day=dt.day - 1)
+        return dt
+
+    def set_date_to_beginning_of_year(dt):
+        """Set date to the beginning of the year."""
+
+        dt = dt.replace(month=1, day=1)
+
         return dt
 
     raw_group_data = yaml.load(
@@ -377,10 +384,26 @@ def _load_parameter_group_from_yaml(
                     )
                     if param in params_last_year:
                         out_params[f"{param}_vorjahr"] = params_last_year[param]
+                elif raw_group_data[param]["access_different_date"] == "jahresanfang":
+                    date_beginning_of_year = set_date_to_beginning_of_year(date)
+                    if date_beginning_of_year == date:
+                        out_params[f"{param}_jahresanfang"] = out_params[param]
+                    else:
+                        params_beginning_of_year = _load_parameter_group_from_yaml(
+                            date_beginning_of_year,
+                            group,
+                            parameters=[param],
+                            yaml_path=yaml_path,
+                        )
+                        if param in params_beginning_of_year:
+                            out_params[
+                                f"{param}_jahresanfang"
+                            ] = params_beginning_of_year[param]
                 else:
                     raise ValueError(
                         "Currently, access_different_date is only implemented for "
-                        "'vorjahr' (last year). "
+                        "'vorjahr' (last year) and "
+                        "'jahresanfang' (beginning of the year). "
                         f"For parameter {param} a different string is specified."
                     )
 
