@@ -1,7 +1,48 @@
 from _gettsim.shared import dates_active
 
 
-def ges_krankenv_beitr_m(  # noqa: PLR0913
+@dates_active(end="2003-03-31", change_name="ges_krankenv_beitr_m")
+def ges_krankenv_beitr_m_vor_midijob(
+    geringfügig_beschäftigt: bool,
+    ges_krankenv_beitr_rente_m: float,
+    ges_krankenv_beitr_selbst_m: float,
+    _ges_krankenv_beitr_reg_beschäftigt_m: float,
+    selbstständig: bool,
+) -> float:
+    """Contribution for each individual to the public health insurance.
+
+    Parameters
+    ----------
+    geringfügig_beschäftigt
+        See :func:`geringfügig_beschäftigt`.
+    ges_krankenv_beitr_rente_m
+        See :func:`ges_krankenv_beitr_rente_m`.
+    ges_krankenv_beitr_selbst_m
+        See :func:`ges_krankenv_beitr_selbst_m`.
+    _ges_krankenv_beitr_reg_beschäftigt_m
+        See :func:`_ges_krankenv_beitr_reg_beschäftigt_m`.
+    selbstständig
+        See basic input variable :ref:`selbstständig <selbstständig>`.
+
+
+    Returns
+    -------
+
+    """
+
+    if selbstständig:
+        out = ges_krankenv_beitr_selbst_m
+    elif geringfügig_beschäftigt:
+        out = 0.0
+    else:
+        out = _ges_krankenv_beitr_reg_beschäftigt_m
+
+    # Add the health insurance contribution for pensions
+    return out + ges_krankenv_beitr_rente_m
+
+
+@dates_active(start="2003-04-01", change_name="ges_krankenv_beitr_m")
+def ges_krankenv_beitr_m_mit_midijob(  # noqa: PLR0913
     geringfügig_beschäftigt: bool,
     ges_krankenv_beitr_rente_m: float,
     ges_krankenv_beitr_selbst_m: float,
@@ -48,7 +89,50 @@ def ges_krankenv_beitr_m(  # noqa: PLR0913
     return out + ges_krankenv_beitr_rente_m
 
 
-def ges_krankenv_beitr_arbeitg_m(
+@dates_active(end="2003-03-31", change_name="ges_krankenv_beitr_arbeitg_m")
+def ges_krankenv_beitr_arbeitg_m_vor_midijob(
+    geringfügig_beschäftigt: bool,
+    bruttolohn_m: float,
+    _ges_krankenv_bruttolohn_m: float,
+    selbstständig: bool,
+    sozialv_beitr_params: dict,
+    _ges_krankenv_beitr_satz_arbeitg: float,
+) -> float:
+    """Contribution of the respective employer to the public health insurance.
+
+    Parameters
+    ----------
+    geringfügig_beschäftigt
+        See :func:`geringfügig_beschäftigt`.
+    _ges_krankenv_bruttolohn_m
+        See :func:`_ges_krankenv_bruttolohn_m`.
+    _ges_krankenv_beitr_satz_arbeitg
+        See :func:`_ges_krankenv_beitr_satz_arbeitg`.
+    bruttolohn_m
+        See basic input variable :ref:`bruttolohn_m <bruttolohn_m>`.
+    selbstständig
+        See basic input variable :ref:`selbstständig <selbstständig>`.
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+
+    Returns
+    -------
+
+    """
+
+    if selbstständig:
+        out = 0.0
+    elif geringfügig_beschäftigt:
+        out = bruttolohn_m * sozialv_beitr_params["ag_abgaben_geringf"]["ges_krankenv"]
+    else:
+        out = _ges_krankenv_bruttolohn_m * _ges_krankenv_beitr_satz_arbeitg
+
+    return out
+
+
+@dates_active(start="2003-04-01", change_name="ges_krankenv_beitr_arbeitg_m")
+def ges_krankenv_beitr_arbeitg_m_mit_midijob(
     geringfügig_beschäftigt: bool,
     in_gleitzone: bool,
     bruttolohn_m: float,
@@ -102,7 +186,9 @@ def ges_krankenv_beitr_arbeitg_m(
     end="2014-12-31",
     change_name="ges_krankenv_zusatzbeitr_satz",
 )
-def ges_krankenv_sonderbeitr_satz(sozialv_beitr_params: dict) -> float:
+def ges_krankenv_zusatzbeitr_satz_from_sonderbeitr_satz(
+    sozialv_beitr_params: dict,
+) -> float:
     """Calculate the top-up rate of the health care insurance.
 
     Parameters
@@ -123,7 +209,9 @@ def ges_krankenv_sonderbeitr_satz(sozialv_beitr_params: dict) -> float:
     start="2015-01-01",
     change_name="ges_krankenv_zusatzbeitr_satz",
 )
-def ges_krankenv_mean_zusatzbeitrag(sozialv_beitr_params: dict) -> float:
+def ges_krankenv_zusatzbeitr_satz_from_mean_zusatzbeitrag(
+    sozialv_beitr_params: dict,
+) -> float:
     """Calculate the top-up rate of the health care insurance.
 
     Parameters
@@ -146,7 +234,7 @@ def ges_krankenv_mean_zusatzbeitrag(sozialv_beitr_params: dict) -> float:
 def ges_krankenv_beitr_satz(
     sozialv_beitr_params: dict,
 ) -> float:
-    """Select contribution rates of employees for health insurance, just a basic split
+    """Contribution rates of employees for health insurance, just a basic split
     between employees and employers. Incorporates regime changes regarding different
     values across insurers (pick "official" mean) and same contribution rate for all.
 
@@ -160,26 +248,53 @@ def ges_krankenv_beitr_satz(
     Beitragssatz for statutory health insurance.
 
     """
-    params = sozialv_beitr_params["beitr_satz"]["ges_krankenv"]
 
+    return sozialv_beitr_params["beitr_satz"]["ges_krankenv"]["mean_allgemein"] / 2
+
+
+@dates_active(
+    end="2005-06-30",
+)
+def _ges_krankenv_beitr_satz_jahresanfang(
+    sozialv_beitr_params: dict,
+) -> float:
+    """Contribution rates of employees for health insurance for the beginning
+    of a year, just a basic split between employees and employers. Incorporates
+    regime changes regarding different values across insurers (pick "official" mean)
+    and same contribution rate for all.
+
+    Parameters
+    ----------
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+    Beitragssatz for statutory health insurance at the begging of the year.
+
+    """
     return (
-        params["allgemein"] if "allgemein" in params else params["mean_allgemein"]
-    ) / 2
+        sozialv_beitr_params["beitr_satz_jahresanfang"]["ges_krankenv"][
+            "mean_allgemein"
+        ]
+        / 2
+    )
 
 
 @dates_active(
     start="2005-07-01",
-    end="2018-12-31",
+    end="2008-12-31",
     change_name="ges_krankenv_beitr_satz",
 )
-def ges_krankenv_beitr_satz_zusatzbeitrag_nur_arbeitn(
+def ges_krankenv_beitr_satz_mean_kassenspezifisch_zusatzbeitrag_nur_arbeitn(
     ges_krankenv_zusatzbeitr_satz: float,
     sozialv_beitr_params: dict,
 ) -> float:
-    """Select contribution rates of employees for health insurance until 2018.
+    """Contribution rates of employees for health insurance.
 
     The contribution rates consists of a general rate (split equally between employers
-    and employees) and a top-up rate which is fully paid by employees.
+    and employees, differs across sickness funds) and a top-up rate, which is fully paid
+    by employees.
 
     Parameters
     ----------
@@ -190,17 +305,120 @@ def ges_krankenv_beitr_satz_zusatzbeitrag_nur_arbeitn(
 
     Returns
     -------
+    Beitragssatz for statutory health insurance.
 
     """
-    params = sozialv_beitr_params["beitr_satz"]["ges_krankenv"]
 
-    # Contribution rates differ between insurance entities until 2008.
-    # We, hence, rely on average contribution rates "mean_allgemein" for these years.
-    allgemeiner_beitrag = (
-        params["allgemein"] if "allgemein" in params else params["mean_allgemein"]
-    )
+    mean_allgemein = sozialv_beitr_params["beitr_satz"]["ges_krankenv"][
+        "mean_allgemein"
+    ]
 
-    return allgemeiner_beitrag / 2 + ges_krankenv_zusatzbeitr_satz
+    return mean_allgemein / 2 + ges_krankenv_zusatzbeitr_satz
+
+
+@dates_active(
+    start="2005-07-01",
+    end="2008-12-31",
+    change_name="_ges_krankenv_beitr_satz_jahresanfang",
+)
+def ges_krankenv_beitr_satz_mean_kassenspezifisch_zusatzbeitrag_nur_arbeitn_jahresanfang(  # noqa: E501
+    ges_krankenv_zusatzbeitr_satz: float,
+    sozialv_beitr_params: dict,
+) -> float:
+    """Contribution rates of employees for health insurance at the beginning of the
+    year.
+
+    The contribution rates consists of a general rate (split equally between employers
+    and employees, differs across sickness funds) and a top-up rate, which is fully paid
+    by employees.
+
+    Parameters
+    ----------
+    ges_krankenv_zusatzbeitr_satz
+        See :func:`ges_krankenv_zusatzbeitr_satz`.
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+    Beitragssatz for statutory health insurance at the beginning of the year.
+
+    """
+
+    mean_allgemein = sozialv_beitr_params["beitr_satz_jahresanfang"]["ges_krankenv"][
+        "mean_allgemein"
+    ]
+
+    return mean_allgemein / 2 + ges_krankenv_zusatzbeitr_satz
+
+
+@dates_active(
+    start="2009-01-01",
+    end="2018-12-31",
+    change_name="ges_krankenv_beitr_satz",
+)
+def ges_krankenv_beitr_satz_einheitlich_zusatzbeitrag_nur_arbeitn(
+    ges_krankenv_zusatzbeitr_satz: float,
+    sozialv_beitr_params: dict,
+) -> float:
+    """Contribution rates of employees for health insurance.
+
+    The contribution rates consists of a general rate (split equally between employers
+    and employees, same for all sickness funds) and a top-up rate, which is fully paid
+    by employees.
+
+    Parameters
+    ----------
+    ges_krankenv_zusatzbeitr_satz
+        See :func:`ges_krankenv_zusatzbeitr_satz`.
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+    Beitragssatz for statutory health insurance.
+
+    """
+
+    allgemein = sozialv_beitr_params["beitr_satz"]["ges_krankenv"]["allgemein"]
+
+    return allgemein / 2 + ges_krankenv_zusatzbeitr_satz
+
+
+@dates_active(
+    start="2009-01-01",
+    end="2018-12-31",
+    change_name="_ges_krankenv_beitr_satz_jahresanfang",
+)
+def ges_krankenv_beitr_satz_einheitlich_zusatzbeitrag_nur_arbeitn_jahresanfang(
+    ges_krankenv_zusatzbeitr_satz: float,
+    sozialv_beitr_params: dict,
+) -> float:
+    """Contribution rates of employees for health insurance at the beginning of the
+    year.
+
+    The contribution rates consists of a general rate (split equally between employers
+    and employees, same for all sickness funds) and a top-up rate, which is fully paid
+    by employees.
+
+    Parameters
+    ----------
+    ges_krankenv_zusatzbeitr_satz
+        See :func:`ges_krankenv_zusatzbeitr_satz`.
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+    Beitragssatz for statutory health insurance at the beginning of the year.
+
+    """
+
+    allgemein = sozialv_beitr_params["beitr_satz_jahresanfang"]["ges_krankenv"][
+        "allgemein"
+    ]
+
+    return allgemein / 2 + ges_krankenv_zusatzbeitr_satz
 
 
 @dates_active(
@@ -211,7 +429,35 @@ def ges_krankenv_beitr_satz_zusatzbeitrag_paritätisch(
     ges_krankenv_zusatzbeitr_satz: float,
     sozialv_beitr_params: dict,
 ) -> float:
-    """Select contribution rates of employees for health insurance since 2019.
+    """Contribution rates of employees for health insurance.
+
+    Zusatzbeitrag is split equally between employers and employees.
+
+    Parameters
+    ----------
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+
+    """
+    allgemeiner_beitr_satz = sozialv_beitr_params["beitr_satz"]["ges_krankenv"][
+        "allgemein"
+    ]
+    return (allgemeiner_beitr_satz + ges_krankenv_zusatzbeitr_satz) / 2
+
+
+@dates_active(
+    start="2019-01-01",
+    change_name="_ges_krankenv_beitr_satz_jahresanfang",
+)
+def ges_krankenv_beitr_satz_zusatzbeitrag_paritätisch_jahresanfang(
+    ges_krankenv_zusatzbeitr_satz: float,
+    sozialv_beitr_params: dict,
+) -> float:
+    """Contribution rates of employees for health insurance at the beginning of the
+    year.
 
     Zusatzbeitrag is now split equally between employers and employees.
 
@@ -224,20 +470,20 @@ def ges_krankenv_beitr_satz_zusatzbeitrag_paritätisch(
     -------
 
     """
-    params = sozialv_beitr_params["beitr_satz"]["ges_krankenv"]
-    allgemeiner_beitrag = params["allgemein"]
-    zusatzbeitrag = ges_krankenv_zusatzbeitr_satz
-    return (allgemeiner_beitrag + zusatzbeitrag) / 2
+    allgemeiner_beitr_satz = sozialv_beitr_params["beitr_satz_jahresanfang"][
+        "ges_krankenv"
+    ]["allgemein"]
+    return (allgemeiner_beitr_satz + ges_krankenv_zusatzbeitr_satz) / 2
 
 
 @dates_active(
-    end="2018-12-31",
+    end="2008-12-31",
     change_name="_ges_krankenv_beitr_satz_arbeitg",
 )
-def _ges_krankenv_beitr_satz_arbeitg_zusatzbeitrag_nur_arbeitn(
+def ges_krankenv_beitr_satz_arbeitg_mean_kassenspezifisch_zusatzbeitrag_nur_arbeitn(
     sozialv_beitr_params: dict,
 ) -> float:
-    """Select contribution rates of employers for health insurance until 2018.
+    """Contribution rates of employers for health insurance. Zusatzbeitrag irrelevant.
 
     Parameters
     ----------
@@ -248,13 +494,85 @@ def _ges_krankenv_beitr_satz_arbeitg_zusatzbeitrag_nur_arbeitn(
     -------
 
     """
-    params = sozialv_beitr_params["beitr_satz"]["ges_krankenv"]
 
-    # Contribution rates differ between insurance entities until 2008.
-    # We, hence, rely on average contribution rates "mean_allgemein".
-    total = params["allgemein"] if "allgemein" in params else params["mean_allgemein"]
+    return sozialv_beitr_params["beitr_satz"]["ges_krankenv"]["mean_allgemein"] / 2
 
-    return total / 2
+
+@dates_active(
+    end="2008-12-31",
+    change_name="_ges_krankenv_beitr_satz_arbeitg_jahresanfang",
+)
+def ges_krankenv_beitr_satz_arbeitg_mean_kassenspezifisch_zusatzbeitrag_nur_arbeitn_jahresanfang(  # noqa: E501
+    sozialv_beitr_params: dict,
+) -> float:
+    """Contribution rates of employers for health insurance. Zusatzbeitrag irrelevant.
+
+    Parameters
+    ----------
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+
+    """
+
+    return (
+        sozialv_beitr_params["beitr_satz_jahresanfang"]["ges_krankenv"][
+            "mean_allgemein"
+        ]
+        / 2
+    )
+
+
+@dates_active(
+    start="2009-01-01",
+    end="2018-12-31",
+    change_name="_ges_krankenv_beitr_satz_arbeitg",
+)
+def ges_krankenv_beitr_satz_arbeitg_einheitlich_zusatzbeitrag_nur_arbeitn(
+    sozialv_beitr_params: dict,
+) -> float:
+    """Contribution rates of employers for health insurance. Uniform contribution rate
+    for all health insurers, Zusatzbeitrag irrelevant.
+
+    Parameters
+    ----------
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+
+    """
+
+    return sozialv_beitr_params["beitr_satz"]["ges_krankenv"]["allgemein"] / 2
+
+
+@dates_active(
+    start="2009-01-01",
+    end="2018-12-31",
+    change_name="_ges_krankenv_beitr_satz_arbeitg_jahresanfang",
+)
+def ges_krankenv_beitr_satz_arbeitg_einheitlich_zusatzbeitrag_nur_arbeitn_jahresanfang(
+    sozialv_beitr_params: dict,
+) -> float:
+    """Contribution rates of employers for health insurance at the beginning of the
+    year. Uniform contribution rate for all health insurers, Zusatzbeitrag irrelevant.
+
+    Parameters
+    ----------
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+
+    """
+
+    return (
+        sozialv_beitr_params["beitr_satz_jahresanfang"]["ges_krankenv"]["allgemein"] / 2
+    )
 
 
 @dates_active(
@@ -264,7 +582,7 @@ def _ges_krankenv_beitr_satz_arbeitg_zusatzbeitrag_nur_arbeitn(
 def _ges_krankenv_beitr_satz_arbeitg_zusatzbeitrag_paritätisch(
     ges_krankenv_beitr_satz: float,
 ) -> float:
-    """Select contribution rates of employers for health insurance since 2019.
+    """Contribution rates of employers for health insurance since 2019.
 
     The full contribution rate is now split equally between employers and employees.
 
@@ -278,6 +596,30 @@ def _ges_krankenv_beitr_satz_arbeitg_zusatzbeitrag_paritätisch(
 
     """
     return ges_krankenv_beitr_satz
+
+
+@dates_active(
+    start="2019-01-01",
+    change_name="_ges_krankenv_beitr_satz_arbeitg_jahresanfang",
+)
+def _ges_krankenv_beitr_satz_arbeitg_zusatzbeitrag_paritätisch_jahresanfang(
+    _ges_krankenv_beitr_satz_jahresanfang: float,
+) -> float:
+    """Contribution rates of employers for health insurance for the beginning
+    of a year since 2019.
+
+    The full contribution rate is now split equally between employers and employees.
+
+    Parameters
+    ----------
+    ges_krankenv_beitr_satz
+        See :func:`ges_krankenv_beitr_satz`.
+
+    Returns
+    -------
+
+    """
+    return _ges_krankenv_beitr_satz_jahresanfang
 
 
 def _ges_krankenv_bruttolohn_m(
@@ -456,6 +798,7 @@ def ges_krankenv_beitr_rente_m(
     return ges_krankenv_beitr_satz * _ges_krankenv_bemessungsgrundlage_rente_m
 
 
+@dates_active(start="2003-04-01")
 def _ges_krankenv_beitr_midijob_sum_arbeitn_arbeitg_m(
     midijob_bemessungsentgelt_m: float,
     ges_krankenv_beitr_satz: float,
@@ -484,6 +827,7 @@ def _ges_krankenv_beitr_midijob_sum_arbeitn_arbeitg_m(
 
 
 @dates_active(
+    start="2003-04-01",
     end="2022-09-30",
     change_name="_ges_krankenv_beitr_midijob_arbeitg_m",
 )
@@ -548,6 +892,7 @@ def _ges_krankenv_beitr_midijob_arbeitg_m_residuum(
 
 
 @dates_active(
+    start="2003-04-01",
     end="2022-09-30",
     change_name="_ges_krankenv_beitr_midijob_arbeitn_m",
 )
