@@ -1,6 +1,7 @@
 from _gettsim.shared import dates_active
 
 
+@dates_active(start="2005-01-01")
 def ges_pflegev_zusatz_kinderlos(
     hat_kinder: bool,
     alter: int,
@@ -8,35 +9,156 @@ def ges_pflegev_zusatz_kinderlos(
 ) -> bool:
     """Whether additional care insurance contribution for childless individuals applies.
 
-    ToDo: Make dependent on year. Current implementation is deliberately ugly.
-
     Parameters
     ----------
     hat_kinder
         See basic input variable :ref:`hat_kinder <hat_kinder>`.
     alter
         See basic input variable :ref:`alter <alter>`.
+    sozialv_beitr_params: dict,
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
 
     Returns
     -------
 
     """
-    if "ges_pflegev_zusatz_kinderlos_mindestalter" in sozialv_beitr_params:
-        altersgrenze = sozialv_beitr_params["ges_pflegev_zusatz_kinderlos_mindestalter"]
-    else:
-        altersgrenze = 9999
-    out = (not hat_kinder) and alter >= altersgrenze
+    mindestalter = sozialv_beitr_params["ges_pflegev_zusatz_kinderlos_mindestalter"]
+    return (not hat_kinder) and alter >= mindestalter
+
+
+@dates_active(
+    start="1995-01-01", end="2004-12-31", change_name="ges_pflegev_beitr_satz"
+)
+def ges_pflegev_beitr_satz_ohne_zusatz_fuer_kinderlose(
+    sozialv_beitr_params: dict,
+) -> float:
+    """Long-term care insurance contribution rate.
+
+    Parameters
+    ----------
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+
+    """
+
+    return sozialv_beitr_params["beitr_satz"]["ges_pflegev"]
+
+
+@dates_active(
+    start="2005-01-01", end="2023-06-30", change_name="ges_pflegev_beitr_satz"
+)
+def ges_pflegev_beitr_satz_zusatz_kinderlos_dummy(
+    ges_pflegev_zusatz_kinderlos: bool,
+    sozialv_beitr_params: dict,
+) -> float:
+    """Long-term care insurance contribution rate.
+
+    Parameters
+    ----------
+    ges_pflegev_zusatz_kinderlos
+        See :func:`ges_pflegev_zusatz_kinderlos`.
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+
+    """
+    out = sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
+
+    # Add additional contribution for childless individuals
+    if ges_pflegev_zusatz_kinderlos:
+        out += sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["zusatz_kinderlos"]
+
     return out
 
 
-def ges_pflegev_beitr_m(  # noqa: PLR0913
+@dates_active(start="2023-07-01", change_name="ges_pflegev_beitr_satz")
+def ges_pflegev_beitr_satz_mit_kinder_abschlag(
+    anz_eig_kind_bis_24: int,
+    ges_pflegev_zusatz_kinderlos: bool,
+    sozialv_beitr_params: dict,
+) -> float:
+    """Care insurance contribution rate.
+    For individuals with children younger than 25 rates are reduced.
+
+    Parameters
+    ----------
+    anz_eig_kind_bis_24: int,
+        See basic input variable :ref:`anz_eig_kind_bis_24 <anz_eig_kind_bis_24>`.
+    ges_pflegev_zusatz_kinderlos
+        See :func:`ges_pflegev_zusatz_kinderlos`.
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+
+    """
+    out = sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
+
+    # Add additional contribution for childless individuals
+    if ges_pflegev_zusatz_kinderlos:
+        out += sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["zusatz_kinderlos"]
+
+    # Reduced contribution for individuals with two or more children under 25
+    if anz_eig_kind_bis_24 >= 2:
+        out -= sozialv_beitr_params["beitr_satz"]["ges_pflegev"][
+            "abschlag_kinder"
+        ] * min(anz_eig_kind_bis_24 - 1, 4)
+
+    return out
+
+
+@dates_active(end="2003-03-31", change_name="ges_pflegev_beitr_m")
+def ges_pflegev_beitr_m_vor_midijob(
+    _ges_pflegev_beitr_reg_beschäftigt_m: float,
+    geringfügig_beschäftigt: bool,
+    ges_pflegev_beitr_rente_m: float,
+    ges_pflegev_beitr_selbst_m: float,
+    selbstständig: bool,
+) -> float:
+    """Contribution for each individual to the public care insurance.
+
+    Parameters
+    ----------
+    geringfügig_beschäftigt
+        See :func:`geringfügig_beschäftigt`.
+    ges_pflegev_beitr_rente_m
+        See :func:`ges_pflegev_beitr_rente_m`.
+    ges_pflegev_beitr_selbst_m
+        See :func:`ges_pflegev_beitr_selbst_m`.
+    ges_pflegev_beitr_regulär_besch_m
+        See :func:`ges_pflegev_beitr_regulär_besch_m`.
+    selbstständig
+        See basic input variable :ref:`selbstständig <selbstständig>`.
+
+    Returns
+    -------
+
+    """
+
+    if selbstständig:
+        out = ges_pflegev_beitr_selbst_m
+    elif geringfügig_beschäftigt:
+        out = 0.0
+    else:
+        out = _ges_pflegev_beitr_reg_beschäftigt_m
+
+    # Add the care insurance contribution for pensions
+    return out + ges_pflegev_beitr_rente_m
+
+
+@dates_active(start="2003-04-01", change_name="ges_pflegev_beitr_m")
+def ges_pflegev_beitr_m_mit_midijob(  # noqa: PLR0913
+    _ges_pflegev_beitr_reg_beschäftigt_m: float,
     geringfügig_beschäftigt: bool,
     ges_pflegev_beitr_rente_m: float,
     ges_pflegev_beitr_selbst_m: float,
     _ges_pflegev_beitr_midijob_arbeitn_m: float,
-    ges_pflegev_zusatz_kinderlos: bool,
-    _ges_krankenv_bruttolohn_m: float,
-    sozialv_beitr_params: dict,
     in_gleitzone: bool,
     selbstständig: bool,
 ) -> float:
@@ -52,12 +174,8 @@ def ges_pflegev_beitr_m(  # noqa: PLR0913
         See :func:`ges_pflegev_beitr_selbst_m`.
     _ges_pflegev_beitr_midijob_arbeitn_m
         See :func:`_ges_pflegev_beitr_midijob_arbeitn_m`.
-    ges_pflegev_zusatz_kinderlos
-        See :func:`ges_pflegev_zusatz_kinderlos`.
-    _ges_krankenv_bruttolohn_m
-        See :func:`_ges_krankenv_bruttolohn_m`.
-    sozialv_beitr_params
-        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+    ges_pflegev_beitr_regulär_besch_m
+        See :func:`ges_pflegev_beitr_regulär_besch_m`.
     in_gleitzone
         See :func:`in_gleitzone`.
     selbstständig
@@ -68,19 +186,6 @@ def ges_pflegev_beitr_m(  # noqa: PLR0913
 
     """
 
-    # Calculate care insurance contributions for regular jobs.
-    beitr_regulär_beschäftigt_m = (
-        _ges_krankenv_bruttolohn_m
-        * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
-    )
-
-    # Add additional contribution for childless individuals
-    if ges_pflegev_zusatz_kinderlos:
-        beitr_regulär_beschäftigt_m += (
-            _ges_krankenv_bruttolohn_m
-            * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["zusatz_kinderlos"]
-        )
-
     if selbstständig:
         out = ges_pflegev_beitr_selbst_m
     elif geringfügig_beschäftigt:
@@ -88,13 +193,74 @@ def ges_pflegev_beitr_m(  # noqa: PLR0913
     elif in_gleitzone:
         out = _ges_pflegev_beitr_midijob_arbeitn_m
     else:
-        out = beitr_regulär_beschäftigt_m
+        out = _ges_pflegev_beitr_reg_beschäftigt_m
 
     # Add the care insurance contribution for pensions
     return out + ges_pflegev_beitr_rente_m
 
 
-def ges_pflegev_beitr_arbeitg_m(
+def _ges_pflegev_beitr_reg_beschäftigt_m(
+    _ges_krankenv_bruttolohn_m: float,
+    ges_pflegev_beitr_satz: float,
+) -> float:
+    """Contribution to the public care insurance for people with regular jobs.
+
+    Parameters
+    ----------
+    _ges_krankenv_bruttolohn_m:
+        See :func:`_ges_krankenv_bruttolohn_m`.
+    ges_pflegev_beitr_satz:
+        See :func:`ges_pflegev_beitr_satz`.
+
+    Returns
+    -------
+
+    """
+
+    beitr_regulär_beschäftigt_m = _ges_krankenv_bruttolohn_m * ges_pflegev_beitr_satz
+
+    return beitr_regulär_beschäftigt_m
+
+
+@dates_active(end="2003-03-31", change_name="ges_pflegev_beitr_arbeitg_m")
+def ges_pflegev_beitr_arbeitg_m_vor_midijob(
+    geringfügig_beschäftigt: bool,
+    _ges_krankenv_bruttolohn_m: float,
+    sozialv_beitr_params: dict,
+    selbstständig: bool,
+) -> float:
+    """Contribution of the respective employer to the public care insurance.
+
+    Parameters
+    ----------
+    geringfügig_beschäftigt
+        See :func:`geringfügig_beschäftigt`.
+    _ges_krankenv_bruttolohn_m
+        See :func:`_ges_krankenv_bruttolohn_m`.
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+    selbstständig
+        See basic input variable :ref:`selbstständig <selbstständig>`.
+
+    Returns
+    -------
+
+    """
+    # Calculate care insurance contributions for regular jobs.
+    beitr_regulär_beschäftigt_m = (
+        _ges_krankenv_bruttolohn_m * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]
+    )
+
+    if selbstständig or geringfügig_beschäftigt:
+        out = 0.0
+    else:
+        out = beitr_regulär_beschäftigt_m
+
+    return out
+
+
+@dates_active(start="2003-04-01", change_name="ges_pflegev_beitr_arbeitg_m")
+def ges_pflegev_beitr_arbeitg_m_mit_midijob(
     geringfügig_beschäftigt: bool,
     _ges_pflegev_beitr_midijob_arbeitg_m: float,
     _ges_krankenv_bruttolohn_m: float,
@@ -129,9 +295,7 @@ def ges_pflegev_beitr_arbeitg_m(
         * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
     )
 
-    if selbstständig:
-        out = 0.0
-    if geringfügig_beschäftigt:
+    if selbstständig or geringfügig_beschäftigt:
         out = 0.0
     elif in_gleitzone:
         out = _ges_pflegev_beitr_midijob_arbeitg_m
@@ -141,24 +305,56 @@ def ges_pflegev_beitr_arbeitg_m(
     return out
 
 
-def ges_pflegev_beitr_selbst_m(
-    ges_pflegev_zusatz_kinderlos: bool,
+@dates_active(
+    start="1995-01-01", end="2004-12-31", change_name="ges_pflegev_beitr_selbst_m"
+)
+def ges_pflegev_beitr_selbst_m_ohne_zusatz_fuer_kinderlose(
     _ges_krankenv_bemessungsgrundlage_eink_selbst: float,
+    ges_pflegev_beitr_satz: float,
+) -> float:
+    """Calculate care insurance contributions for self-employed individuals.
+
+    Self-employed pay the full contribution (employer + employee), which is either
+    assessed on their self-employement income or 3/4 of the 'Bezugsgröße'
+
+    Parameters
+    ----------
+
+    _ges_krankenv_bemessungsgrundlage_eink_selbst
+        See :func:`_ges_krankenv_bemessungsgrundlage_eink_selbst`.
+
+    ges_pflegev_beitr_satz
+        See :func:`ges_pflegev_beitr_satz`.
+
+    Returns
+    -------
+    Monthly care insurance contributions for self employed income.
+
+    """
+    out = _ges_krankenv_bemessungsgrundlage_eink_selbst * (ges_pflegev_beitr_satz * 2)
+
+    return out
+
+
+@dates_active(start="2005-01-01", change_name="ges_pflegev_beitr_selbst_m")
+def ges_pflegev_beitr_selbst_m_zusatz_kinderlos_dummy(
+    _ges_krankenv_bemessungsgrundlage_eink_selbst: float,
+    ges_pflegev_beitr_satz: float,
     sozialv_beitr_params: dict,
 ) -> float:
     """Calculate care insurance contributions for self-employed individuals.
 
-    Self-employed pay the full
-    contribution (employer + employee), which is either assessed on their
-    self-employement income or 3/4 of the 'Bezugsgröße'
+    Self-employed pay the full contribution (employer + employee), which is either
+    assessed on their self-employement income or 3/4 of the 'Bezugsgröße'
 
     Parameters
     ----------
-    ges_pflegev_zusatz_kinderlos
-        See :func:`ges_pflegev_zusatz_kinderlos`.
 
     _ges_krankenv_bemessungsgrundlage_eink_selbst
         See :func:`_ges_krankenv_bemessungsgrundlage_eink_selbst`.
+
+    ges_pflegev_beitr_satz
+        See :func:`ges_pflegev_beitr_satz`.
 
     sozialv_beitr_params
         See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
@@ -168,35 +364,59 @@ def ges_pflegev_beitr_selbst_m(
     Monthly care insurance contributions for self employed income.
 
     """
-    out = (
-        _ges_krankenv_bemessungsgrundlage_eink_selbst
-        * 2
-        * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
+    out = _ges_krankenv_bemessungsgrundlage_eink_selbst * (
+        ges_pflegev_beitr_satz
+        + sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
     )
-
-    # Add additional contribution for childless individuals
-    if ges_pflegev_zusatz_kinderlos:
-        out += (
-            _ges_krankenv_bemessungsgrundlage_eink_selbst
-            * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["zusatz_kinderlos"]
-        )
 
     return out
 
 
-def ges_pflegev_beitr_rente_m(
-    ges_pflegev_zusatz_kinderlos: bool,
+@dates_active(
+    start="1995-01-01", end="2004-12-31", change_name="ges_pflegev_beitr_rente_m"
+)
+def ges_pflegev_beitr_rente_m_ohne_zusatz_fuer_kinderlose(
     _ges_krankenv_bemessungsgrundlage_rente_m: float,
+    ges_pflegev_beitr_satz: float,
+) -> float:
+    """Calculating the contribution to health insurance for pension income.
+
+    Pensioners pay twice the contribution of employees.
+
+    Parameters
+    ----------
+    _ges_krankenv_bemessungsgrundlage_rente_m
+        See :func:`_ges_krankenv_bemessungsgrundlage_rente_m`.
+    ges_pflegev_beitr_satz
+        See :func:`ges_pflegev_beitr_satz`.
+
+    Returns
+    -------
+    Monthly health insurance contributions for pension income.
+
+    """
+    out = _ges_krankenv_bemessungsgrundlage_rente_m * (ges_pflegev_beitr_satz * 2)
+
+    return out
+
+
+@dates_active(start="2005-01-01", change_name="ges_pflegev_beitr_rente_m")
+def ges_pflegev_beitr_rente_m_zusatz_kinderlos_dummy(
+    _ges_krankenv_bemessungsgrundlage_rente_m: float,
+    ges_pflegev_beitr_satz: float,
     sozialv_beitr_params: dict,
 ) -> float:
     """Calculating the contribution to health insurance for pension income.
 
+    Pensioners pay twice the contribution of employees, but only once the additional
+    charge for childless individuals.
+
     Parameters
     ----------
-    ges_pflegev_zusatz_kinderlos
-        See :func:`ges_pflegev_zusatz_kinderlos`.
     _ges_krankenv_bemessungsgrundlage_rente_m
         See :func:`_ges_krankenv_bemessungsgrundlage_rente_m`.
+    ges_pflegev_beitr_satz
+        See :func:`ges_pflegev_beitr_satz`.
     sozialv_beitr_params
         See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
 
@@ -205,26 +425,19 @@ def ges_pflegev_beitr_rente_m(
     Monthly health insurance contributions for pension income.
 
     """
-    out = (
-        _ges_krankenv_bemessungsgrundlage_rente_m
-        * 2
-        * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
+    out = _ges_krankenv_bemessungsgrundlage_rente_m * (
+        ges_pflegev_beitr_satz
+        + sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
     )
-
-    # Add additional contribution for childless individuals
-    if ges_pflegev_zusatz_kinderlos:
-        out += (
-            _ges_krankenv_bemessungsgrundlage_rente_m
-            * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["zusatz_kinderlos"]
-        )
 
     return out
 
 
+@dates_active(start="2003-04-01")
 def _ges_pflegev_beitr_midijob_sum_arbeitn_arbeitg_m(
     midijob_bemessungsentgelt_m: float,
+    ges_pflegev_beitr_satz: float,
     sozialv_beitr_params: dict,
-    ges_pflegev_zusatz_kinderlos: bool,
 ) -> float:
     """Sum of employee and employer long-term care insurance contributions.
 
@@ -232,8 +445,8 @@ def _ges_pflegev_beitr_midijob_sum_arbeitn_arbeitg_m(
     ----------
     midijob_bemessungsentgelt_m
         See :func:`midijob_bemessungsentgelt_m`.
-    ges_pflegev_zusatz_kinderlos
-        See :func:`ges_pflegev_zusatz_kinderlos`.
+    ges_pflegev_beitr_satz
+        See :func:`ges_pflegev_beitr_satz`.
     sozialv_beitr_params
         See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
 
@@ -242,18 +455,10 @@ def _ges_pflegev_beitr_midijob_sum_arbeitn_arbeitg_m(
 
     """
 
-    gesamtbeitrag_midijob_m = (
-        midijob_bemessungsentgelt_m
-        * 2
-        * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
+    gesamtbeitrag_midijob_m = midijob_bemessungsentgelt_m * (
+        ges_pflegev_beitr_satz
+        + sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
     )
-
-    # Add additional contribution for childless individuals
-    if ges_pflegev_zusatz_kinderlos:
-        gesamtbeitrag_midijob_m += (
-            midijob_bemessungsentgelt_m
-            * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["zusatz_kinderlos"]
-        )
 
     return gesamtbeitrag_midijob_m
 
@@ -339,7 +544,11 @@ def _ges_pflegev_beitr_midijob_arbeitn_m_residuum(
     return out
 
 
-@dates_active(start="2022-10-01", change_name="_ges_pflegev_beitr_midijob_arbeitn_m")
+@dates_active(
+    start="2022-10-01",
+    end="2023-06-30",
+    change_name="_ges_pflegev_beitr_midijob_arbeitn_m",
+)
 def _ges_pflegev_beitr_midijob_arbeitn_m_anteil_beitragspfl_einnahme(
     ges_pflegev_zusatz_kinderlos: bool,
     _midijob_beitragspfl_einnahme_arbeitn_m: float,
@@ -368,6 +577,55 @@ def _ges_pflegev_beitr_midijob_arbeitn_m_anteil_beitragspfl_einnahme(
         _midijob_beitragspfl_einnahme_arbeitn_m
         * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
     )
+
+    # Add additional contribution for childless individuals
+    if ges_pflegev_zusatz_kinderlos:
+        an_beitr_midijob_m += (
+            midijob_bemessungsentgelt_m
+            * sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["zusatz_kinderlos"]
+        )
+
+    return an_beitr_midijob_m
+
+
+@dates_active(start="2023-07-01", change_name="_ges_pflegev_beitr_midijob_arbeitn_m")
+def _ges_pflegev_beitr_midijob_arbeitn_m_anteil_mit_kinder_abschlag(
+    anz_eig_kind_bis_24: int,
+    ges_pflegev_zusatz_kinderlos: bool,
+    _midijob_beitragspfl_einnahme_arbeitn_m: float,
+    midijob_bemessungsentgelt_m: float,
+    sozialv_beitr_params: dict,
+) -> float:
+    """Calculating the employee care insurance contribution since October 2022.
+
+    Parameters
+    ----------
+    anz_eig_kind_bis_24
+        See basic input variable :ref:`anz_eig_kind_bis_24 <anz_eig_kind_bis_24>`.
+    ges_pflegev_zusatz_kinderlos
+        See :func:`ges_pflegev_zusatz_kinderlos`.
+    midijob_bemessungsentgelt_m
+        See :func:`midijob_bemessungsentgelt_m`.
+    _midijob_beitragspfl_einnahme_arbeitn_m
+        See :func:`_midijob_beitragspfl_einnahme_arbeitn_m`.
+    sozialv_beitr_params
+        See params documentation :ref:`sozialv_beitr_params <sozialv_beitr_params>`.
+
+    Returns
+    -------
+
+    """
+    # Calculate the employee care insurance rate
+    ges_pflegev_rate = sozialv_beitr_params["beitr_satz"]["ges_pflegev"]["standard"]
+
+    # Reduced contribution for individuals with two or more children under 25
+    if anz_eig_kind_bis_24 >= 2:
+        ges_pflegev_rate -= sozialv_beitr_params["beitr_satz"]["ges_pflegev"][
+            "abschlag_kinder"
+        ] * min(anz_eig_kind_bis_24 - 1, 4)
+
+    # Calculate the employee care insurance contribution
+    an_beitr_midijob_m = _midijob_beitragspfl_einnahme_arbeitn_m * ges_pflegev_rate
 
     # Add additional contribution for childless individuals
     if ges_pflegev_zusatz_kinderlos:
