@@ -38,11 +38,9 @@ def arbeitsl_geld_2_eink_m(  # noqa: PLR0913
     Income of a person by unemployment insurance.
 
     """
-    # ToDo: Implement deduction of child income including allowance of 100 EUR.
     if kind:
-        # Do not substract income tax as long as children are still part of the tax
-        # unit of their parents
-        # ToDo: Rewrite once children are in a separate tax unit
+        # TODO (@hmgaudecker): Rewrite once groupings are implemented.
+        # https://github.com/iza-institute-of-labor-economics/gettsim/pull/601
         out = (
             arbeitsl_geld_2_bruttoeink_m
             - sozialv_beitr_m
@@ -200,19 +198,25 @@ def arbeitsl_geld_2_eink_anr_frei_m_basierend_auf_nettoquote(
 @dates_active(start="2005-10-01")
 def arbeitsl_geld_2_eink_anr_frei_m(
     bruttolohn_m: float,
-    anz_kinder_hh: int,
+    eink_selbst_m: float,
+    anz_kinder_bis_17_hh: int,
     arbeitsl_geld_2_params: dict,
 ) -> float:
     """Calculate share of income, which remains to the individual since 10/2005.
 
     Note: Since 2023, Arbeitslosengeld 2 is referred to as Bürgergeld.
+    Sozialgesetzbuch (SGB) Zweites Buch (II) - Bürgergeld, Grundsicherung für
+    Arbeitsuchende. SGB II §11b Abs 3
+    https://www.gesetze-im-internet.de/sgb_2/__11b.html
 
     Parameters
     ----------
     bruttolohn_m
         See basic input variable :ref:`bruttolohn_m <bruttolohn_m>`.
-    anz_kinder_hh
-        See :func:`anz_kinder_hh`.
+    eink_selbst_m
+        See basic input variable :ref:`eink_selbst_m <eink_selbst_m>`.
+    anz_kinder_bis_17_hh
+        See :func:`anz_kinder_bis_17_hh`.
     arbeitsl_geld_2_params
         See params documentation :ref:`arbeitsl_geld_2_params <arbeitsl_geld_2_params>`.
 
@@ -220,10 +224,14 @@ def arbeitsl_geld_2_eink_anr_frei_m(
     -------
 
     """
+    # Beneficiaries who live with a minor child in a group home or who have a minor
+    # child have slightly different thresholds. We currently do not consider the second
+    # condition.
+    eink_erwerbstätigkeit = bruttolohn_m + eink_selbst_m
 
-    if anz_kinder_hh > 0:
+    if anz_kinder_bis_17_hh > 0:
         out = piecewise_polynomial(
-            x=bruttolohn_m,
+            x=eink_erwerbstätigkeit,
             thresholds=arbeitsl_geld_2_params["eink_anr_frei_kinder"]["thresholds"],
             rates=arbeitsl_geld_2_params["eink_anr_frei_kinder"]["rates"],
             intercepts_at_lower_thresholds=arbeitsl_geld_2_params[
@@ -232,7 +240,7 @@ def arbeitsl_geld_2_eink_anr_frei_m(
         )
     else:
         out = piecewise_polynomial(
-            x=bruttolohn_m,
+            x=eink_erwerbstätigkeit,
             thresholds=arbeitsl_geld_2_params["eink_anr_frei"]["thresholds"],
             rates=arbeitsl_geld_2_params["eink_anr_frei"]["rates"],
             intercepts_at_lower_thresholds=arbeitsl_geld_2_params["eink_anr_frei"][
