@@ -173,9 +173,8 @@ def elterngeld_kind(
 
     """
     current_year = elterngeld_params["datum"].astype("datetime64[Y]").astype(int) + 1970
-    out = (
-        current_year - geburtsjahr
-        < list(elterngeld_params["geschw_bonus_altersgrenzen_kinder"].keys())[0]
+    out = current_year - geburtsjahr < next(
+        iter(elterngeld_params["geschw_bonus_altersgrenzen_kinder"].keys())
     )
     return out
 
@@ -229,10 +228,11 @@ def elterngeld_geschw_bonus_anspruch(
 
     """
     if elternzeit_anspruch:
-        # ToDo: Should this be >=? Reference (§ 2 (2) BEEG) is not completely clear
         out = (
             elterngeld_kind_hh
-            == list(elterngeld_params["geschw_bonus_altersgrenzen_kinder"].values())[0]
+            == next(
+                iter(elterngeld_params["geschw_bonus_altersgrenzen_kinder"].values())
+            )
         ) or (
             elterngeld_vorschulkind_hh
             >= list(elterngeld_params["geschw_bonus_altersgrenzen_kinder"].values())[1]
