@@ -23,10 +23,7 @@ def test_arbeitsl_geld(
     )
 
     result = compute_taxes_and_transfers(
-        data=df,
-        params=policy_params,
-        functions=policy_functions,
-        targets=column,
+        data=df, params=policy_params, functions=policy_functions, targets=column
     )
 
     # to prevent errors from rounding, allow deviations after the 3rd digit.
