@@ -14,18 +14,6 @@ from pandas.testing import assert_series_equal
 from _gettsim_tests._helpers import cached_set_up_policy_environment
 from _gettsim_tests._policy_test_utils import PolicyTestData, load_policy_test_data
 
-OVERRIDE_COLS = [
-    "arbeitsl_geld_m",
-    "soli_st_y_tu",
-    "kindergeld_m_hh",
-    "unterhaltsvors_m",
-    "elterngeld_m",
-    "eink_st_y_tu",
-    "sozialv_beitr_m",
-    "sum_ges_rente_priv_rente_m",
-    "wohngeld_vor_vermög_check_m_hh",
-]
-
 data = load_policy_test_data("arbeitsl_geld_2")
 
 
@@ -44,11 +32,7 @@ def test_arbeitsl_geld_2(
     )
 
     result = compute_taxes_and_transfers(
-        data=df,
-        params=policy_params,
-        functions=policy_functions,
-        targets=column,
-        columns_overriding_functions=OVERRIDE_COLS,
+        data=df, params=policy_params, functions=policy_functions, targets=column
     )
 
     if column in [
