@@ -5,15 +5,6 @@ from pandas.testing import assert_series_equal
 from _gettsim_tests._helpers import cached_set_up_policy_environment
 from _gettsim_tests._policy_test_utils import PolicyTestData, load_policy_test_data
 
-OVERRIDE_COLS = [
-    "_kinderzuschl_vor_vermög_check_m_tu",
-    "wohngeld_vor_vermög_check_m_hh",
-    "arbeitsl_geld_2_regelbedarf_m_hh",
-    "kindergeld_m_hh",
-    "unterhaltsvors_m_hh",
-    "arbeitsl_geld_2_eink_m_hh",
-]
-
 data = load_policy_test_data("benefit_checks")
 
 
@@ -32,11 +23,7 @@ def test_benefit_checks(
     )
 
     result = compute_taxes_and_transfers(
-        data=df,
-        params=policy_params,
-        functions=policy_functions,
-        targets=column,
-        columns_overriding_functions=OVERRIDE_COLS,
+        data=df, params=policy_params, functions=policy_functions, targets=column
     )
 
     assert_series_equal(
