@@ -5,12 +5,6 @@ from pandas.testing import assert_series_equal
 from _gettsim_tests._helpers import cached_set_up_policy_environment
 from _gettsim_tests._policy_test_utils import PolicyTestData, load_policy_test_data
 
-OVERRIDE_COLS = [
-    "_zu_verst_eink_ohne_kinderfreib_y_tu",
-    "_zu_verst_eink_mit_kinderfreib_y_tu",
-    "kapitaleink_brutto_y",
-]
-
 data = load_policy_test_data("eink_st")
 
 
@@ -38,11 +32,7 @@ def test_eink_st(
     )
 
     result = compute_taxes_and_transfers(
-        data=df,
-        params=policy_params,
-        functions=policy_functions,
-        targets=column,
-        columns_overriding_functions=OVERRIDE_COLS,
+        data=df, params=policy_params, functions=policy_functions, targets=column
     )
 
     assert_series_equal(
