@@ -5,13 +5,6 @@ from pandas.testing import assert_series_equal
 from _gettsim_tests._helpers import cached_set_up_policy_environment
 from _gettsim_tests._policy_test_utils import PolicyTestData, load_policy_test_data
 
-OVERRIDE_COLS = [
-    "soli_st_y_tu",
-    "sozialv_beitr_m",
-    "eink_st_y_tu",
-]
-
-
 data = load_policy_test_data("elterngeld")
 
 
@@ -41,11 +34,7 @@ def test_elterngeld(
     df["eink_st_y_tu"] = df["eink_st_m"].groupby(df["tu_id"]).transform("sum") * 12
 
     result = compute_taxes_and_transfers(
-        data=df,
-        params=policy_params,
-        functions=policy_functions,
-        targets=column,
-        columns_overriding_functions=OVERRIDE_COLS,
+        data=df, params=policy_params, functions=policy_functions, targets=column
     )
 
     assert_series_equal(
