@@ -1,9 +1,14 @@
 from __future__ import annotations
 
 import textwrap
+from typing import TYPE_CHECKING
 
+import pytest
 from _gettsim.config import RESOURCE_DIR
-from _gettsim.functions_loader import _load_functions
+from _gettsim.functions_loader import _create_derived_functions, _load_functions
+
+if TYPE_CHECKING:
+    from collections.abc import Callable
 
 
 def func():
@@ -46,7 +51,27 @@ def test_special_attribute_module_is_set(tmp_path):
 
 
 def test_special_attribute_module_is_set_for_internal_functions():
-    out = _load_functions("_gettsim.social_insurance_contributions.eink_grenzen")
-    function = out[list(out)[0]]
-
+    a_few_functions = _load_functions(
+        "_gettsim.social_insurance_contributions.eink_grenzen"
+    )
+    function = next(iter(a_few_functions.values()))
     assert function.__module__ == "_gettsim.social_insurance_contributions.eink_grenzen"
+
+
+@pytest.mark.parametrize(
+    ("functions", "targets"),
+    [
+        ({"foo_y": lambda: 1}, ["foo_d_hh"]),
+        ({"foo_y": lambda: 1}, ["foo_d", "foo_d_hh"]),
+    ],
+)
+def test_create_derived_functions(
+    functions: dict[str, Callable], targets: list[str]
+) -> None:
+    time_conversion_functions, aggregation_functions = _create_derived_functions(
+        functions, targets, [], {}
+    )
+    derived_functions = {**time_conversion_functions, **aggregation_functions}
+
+    for name in targets:
+        assert name in derived_functions
