@@ -5,17 +5,6 @@ from pandas.testing import assert_series_equal
 from _gettsim_tests._helpers import cached_set_up_policy_environment
 from _gettsim_tests._policy_test_utils import PolicyTestData, load_policy_test_data
 
-OVERRIDE_COLS = [
-    "eink_st_y_tu",
-    "soli_st_y_tu",
-    "sozialv_beitr_m",
-    "kindergeld_m_hh",
-    "unterhaltsvors_m",
-    "elterngeld_m",
-    "ges_rente_m",
-    "arbeitsl_geld_m",
-]
-
 data = load_policy_test_data("grunds_im_alter")
 
 
@@ -34,11 +23,7 @@ def test_grunds_im_alter(
     )
 
     result = compute_taxes_and_transfers(
-        data=df,
-        params=policy_params,
-        functions=policy_functions,
-        targets=column,
-        columns_overriding_functions=OVERRIDE_COLS,
+        data=df, params=policy_params, functions=policy_functions, targets=column
     )
 
     assert_series_equal(
