@@ -5,13 +5,6 @@ from pandas.testing import assert_series_equal
 from _gettsim_tests._helpers import cached_set_up_policy_environment
 from _gettsim_tests._policy_test_utils import PolicyTestData, load_policy_test_data
 
-OVERRIDE_COLS = [
-    "ges_krankenv_beitr_m",
-    "arbeitsl_v_beitr_m",
-    "ges_pflegev_beitr_m",
-    "ges_rentenv_beitr_m",
-]
-
 data = load_policy_test_data("vorsorgeaufw")
 
 
@@ -30,11 +23,7 @@ def test_vorsorgeaufw(
     )
 
     result = compute_taxes_and_transfers(
-        data=df,
-        params=policy_params,
-        functions=policy_functions,
-        targets=column,
-        columns_overriding_functions=OVERRIDE_COLS,
+        data=df, params=policy_params, functions=policy_functions, targets=column
     )
 
     assert_series_equal(
